Update TestVerbalizer.java
The path between Billgates and UnitedStates is meant to be connected by property birthplace or contry. It was working a while ago. But recently the test is failing. I assume that some data got changed (mayb in latest dbpedia). Billgates birthplace is seattle, If i change the object to Seattle it works fine. I am not sure whether this is the correct fix, but for time being making this fix.